fix: replace GLuint with uint32_t for compatibility with driver abstraction
Panel hierarchy, project browser, and inspector now use uint32_t for
texture handles instead of GLuint, matching the SDK's kguimanager API
change. This is a mechanical rename — both types are unsigned int on
all platforms.

Files: panel_hierarchy.h, panel_hierarchy.cpp, panel_project.h,
       panel_project.cpp, panel_inspector.cpp

+++ src/panel_project.cpp
@@ -97,7 +97,7 @@ ImTextureRef PanelProject::getThumbnailIcon(const kString& uuid, ImTextureRef de
 			thumbPath.string(), thumbKey, kTextureFormat::TEX_FORMAT_RGBA, false);
 		if (tex && tex->getTextureID() != 0)
 		{
-			GLuint glId = (GLuint)tex->getTextureID();
+			uint32_t glId = tex->getTextureID();
 			ImTextureRef ref = (ImTextureRef)(intptr_t)glId;
 			thumbnailCache[uuid] = { glId, ref };
 			return ref;
